+++ domains/certificates/Query.API/API.AppTests/ApiTests.cs
@@ -1,19 +1,27 @@
 using System;
-using System.Linq;
 using System.Net;
 using System.Net.Http.Json;
 using System.Threading.Tasks;
+using API.AppTests.Extensions;
 using API.AppTests.Infrastructure;
+using API.MasterDataService;
 using API.Query.API.ApiModels;
+using CertificateEvents.Primitives;
+using FluentAssertions;
+using IntegrationEvents;
 using Xunit;
 
 namespace API.AppTests;
 
-public class ApiTests : IClassFixture<QueryApiWebApplicationFactory>
+public class ApiTests : IClassFixture<QueryApiWebApplicationFactory>, IClassFixture<MartenDbContainer>
 {
     private readonly QueryApiWebApplicationFactory factory;
 
-    public ApiTests(QueryApiWebApplicationFactory factory) => this.factory = factory;
+    public ApiTests(QueryApiWebApplicationFactory factory, MartenDbContainer martenDbContainer)
+    {
+        this.factory = factory;
+        this.factory.MartenConnectionString = martenDbContainer.ConnectionString;
+    }
 
     [Fact]
     public async Task GetList_UnauthenticatedUser_ReturnsUnauthorized()
@@ -24,17 +32,44 @@ public class ApiTests : IClassFixture<QueryApiWebApplicationFactory>
         Assert.Equal(HttpStatusCode.Unauthorized, certificatesResponse.StatusCode);
     }
 
-    [Fact(Skip = "Skip until postgres with MartenDB can be started by the factory, see https://github.com/Energinet-DataHub/energy-origin-issues/issues/1023")]
-    public async Task GetList_AuthenticatedUser_ReturnsCertificates()
+    [Fact]
+    public async Task GetList_NoCertificates_ReturnsNoContent()
     {
-        var client = factory.CreateAuthenticatedClient(Guid.NewGuid().ToString());
-        var certificatesResponse = await client.GetFromJsonAsync<CertificateList>("certificates");
+        var subject = Guid.NewGuid().ToString();
 
-        const int numberOfMeteringPoints = 3;
-        const int numberOfHours = 24;
-        const int expected = numberOfMeteringPoints * numberOfHours;
+        using var client = factory.CreateAuthenticatedClient(subject);
 
-        Assert.NotNull(certificatesResponse);
-        Assert.Equal(expected, certificatesResponse.Result.Count());
+        var response = await client.GetAsync("certificates");
+
+        response.StatusCode.Should().Be(HttpStatusCode.NoContent);
+    }
+
+    [Fact]
+    public async Task GetList_MeasurementAddedToBus_ReturnsList()
+    {
+        var subject = Guid.NewGuid().ToString();
+
+        factory.AddMasterData(new MasterData(
+            GSRN: "GSRN",
+            GridArea: "GridArea",
+            Type: MeteringPointType.Production,
+            Technology: new Technology("foo", "bar"),
+            MeteringPointOwner: subject,
+            MeteringPointOnboardedStartDate: DateTimeOffset.Now.AddYears(-1)));
+
+        var bus = factory.GetMassTransitBus();
+
+        await bus.Publish(new EnergyMeasuredIntegrationEvent(
+            GSRN: "GSRN",
+            DateFrom: DateTimeOffset.Now.AddHours(-1).ToUnixTimeSeconds(),
+            DateTo: DateTimeOffset.Now.ToUnixTimeSeconds(),
+            Quantity: 42,
+            Quality: MeasurementQuality.Measured));
+
+        using var client = factory.CreateAuthenticatedClient(subject);
+
+        var apiResponse = await client.RepeatedlyGetUntil("certificates", res => res.StatusCode == HttpStatusCode.OK);
+        var certificateList = await apiResponse.Content.ReadFromJsonAsync<CertificateList>();
+        certificateList?.Result.Should().HaveCount(1);
     }
 }
